Crate repo_backup

Structs

Gitlab
A Gitlab client which can be used as a Provider.

GitlabConfig

Repository

Enums

DownloadInfo
Information used to download a Repository.

FetchError
An error that may occur while fetching download targets with Provider::repositories().

Traits

Provider
A source of Repositories.